Unify pattern handling across syntax and interpreter
