+++ src/syntax/src/lib.rs
@@ -50,6 +50,7 @@ pub mod formatter;
 pub mod mika;
 pub mod grammar;
 pub mod literals;
+pub mod patterns;
 pub mod state_machines;
 pub mod functions;
 pub mod repl;
@@ -67,6 +68,7 @@ pub use crate::formatter::*;
 pub use crate::mika::*;
 pub use crate::grammar::*;
 pub use crate::literals::*;
+pub use crate::patterns::*;
 pub use crate::state_machines::*;
 pub use crate::functions::*;
 pub use crate::repl::*;
